+++ miles_plugins/mbridge/__init__.py
@@ -1,5 +1,5 @@
 from .deepseek_v32 import DeepseekV32Bridge
-from .deepseekv4 import DeepseekV4Bridge
+from .deepseek_v4 import DeepseekV4Bridge
 from .glm4 import GLM4Bridge
 from .glm4moe import GLM4MoEBridge
 from .glm4moe_lite import GLM4MoELiteBridge
